Refactor into multiple plugins
To support building libraries and deployables in different ways, it
makes sense to split up our plugin into two: li-python-library and
li-python-deployable. Each plugin knows how to build and emit its own
artifacts. In the case of li-python-library, that is an sdist. In the
case of li-python-deployable, that is a pex.

More work needs to go into this, but as of now, we can build libraries
and deployable applications, we can run a deployable with `manage.py`,
we cannot run a deployable with `mint deploy`, and we can build
libraries.

RB=456647
R=zpetkovi,rsanders
A=dsully