Fix mppy resolution backwards compatibility
This change is a part in what we need to do to maintain backwards
compatibility with mppy's resolution engine.

This change dumps the transitive dependency tree into a
requirements.txt-like file called mppy-pinned.txt. This file can be used
by the distgradle distclass to express transitive dependencies to a
non-pygradle based build.

RB=470059
R=zpetkovi,rsanders,dsully
A=dsully