Fix pythonAppDef null pointer exception
This was a miss from a previous refactor that simplified how we handle
the pythonSettings extension.

RB=485355
R=zpetkovi,broberts,xachen,sfaber,dsully

ACLOVERRIDE - this is breaking folks using pygradle with cfg2. I've
emailed the reviewers to take a look at this after a build so we can
unblock people. I've also tested this on a example-pygradle-webapp
locally.