Use a per build pex cache
The pex cache is used for speeding up dependency look ups at build time.
If you don't specify a pex cache, it'll default to ~/.pex/build. This
cache works wonderfully for dependencies with published versions, but
not so well for projects that are locally being developed and thus don't
have idempotent version numbers.

To accomodate for this unfortunate use of caching, we can specify a pex
cache on a per build basis by pointing it at our build directory. We
make our plugin smart enough to blow this cache away per build. In doing
this, we make sure to never run into some of the confusing issues we've
seen with ~/.pex (+ptharani, +dbuthay).

RB=485389
R=zpetkovi,broberts,xachen,dsully,sfaber,dbuthay,ptharani
A=dbuthay,dsully