Implement thin deployables
This change adds thin deployables, implemented using pex and shell
wrapper scripts. By using a thin bash wrapper script we primarily get
the ability to set environment variables that can influence where the
pex cache goes. In doing this, we can solve some of the annoying ~/.pex
caching issues that people have been having.

This change requires a bit of refactoring. I had to move a few closures
into the base plugin as methods, had to add a few items to the
components class, etc.

In moving things to the components class I've noticed that now the build
directory is nested in the subproject now.  That means the deployable
directory and much more shows up in root/subproject/build instead of
root/build. I'm not sure if this is correct or not. Thoughts welcome.

RB=488737
R=zpetkovi,broberts,sfaber,xachen,dsully,lcarvalh
A=dsully