Always rebuild pex files
We've observed that even when a project changes, the pex files are not
regenerated after the first build. This is due to the buildPex task
declaring its outputs to be the deployableBinDir and nothing else.

To solve this problem this change regenerates the pex files every time.
This shouldn't be prohibitively expensive because pex generation is
quick.

We cannot easily use inputs/outputs to solve this problem because
inputs/outputs must be generated at configuration time. We do not know
the inputs/outputs at configuration time because we rely on a method,
collectEntryPoints, that must be run at execution time.

Unrelated to this change: I've refactored the wheel building code into a
method so that we get type checking, and did some logging cleaning up.

RB=493600
R=zpetkovi,broberts,sfaber,xachen,dsully,lcarvalh
A=zpetkovi