Change the deployable extension to tar.gz.
Now that we fixed the docs, we have to artifacts with tar.gz extension,
while the deployable itself has tgz.
There seems to be a bug in lipy-artifactory resolution where it doesn't
look at the classifier (-docs-html), finds the first artifact (a doc
bundle) and assumes that everything has that extension. Then it fails to
find tgz artifact (deployable). This will fix the issue by making the
extension the same. We should still file a ticket against
lipy-artifactory to fix this resolution process.

RB=494028
R=xachen
A=xachen