Fix the packageInfoFromPath method
This method was buggy and didn't gracefully handle packages that had
trailing information after the package version. Because of this,
packages like watchdog-0.0.1-linkedin1 were handled incorrectly with
this code.

This change fixes up the packageInfoFromPath method by i) making it a
method to get type checking goodness and ii) changing logic to use regex
instead of confusing string splitting.

RB=496848
R=zpetkovi,broberts,xachen,dsully,sfaber,dholman
A=broberts