Refactor pipInstall into method
Several changes happen in this commit:

i) Refactor pipInstall from a closure inlined in a plugin to a top level
method that has type checking, documentation, etc. During refactor, we
deleted the broken parallelPip implementation. We need to think more
about this and investigate why native code fails to be installable in
parallel. Until then, we're going to clean up the code a bit. We can add
this bad later if we get it to work, I promise.

ii) Slurp all output so that we can display a pretty log message instead
of a verbose (and not very helpful) log message from pip. Only on error
will the original message be displayed to the end user. This should
address a handful of issues that have come up that ask, "Why did my
build fail," that we answer with, "Run ``ligradle -i build`` then run
the exact command that failed."

iii) Add some flair to our build! This uses ASCII escape codes and
unicode to make the build output a little bit prettier for TTY's that
support it. I have it off by default while I do futher testing on Hudson
and other environments that have different TTY than my own.

RB=499256
R=zpetkovi,broberts,xachen,sfaber,dsully
A=broberts,dsully