The thin wrapper should exec pex.
Otherwise, there are two process and two PIDs.
When trying to `control stop` the app, the PID of a wrapper is sent the
TERM signal instead of the PID of the pex.
The exec ensures that there's only one PID, just like with fat pex.

RB=499505
BUGS=TOOLS-63231
R=sholsapp
A=sholsapp,dsully