Use 'pythonEnvironmentDistgradle' for '...python setup.py entrypoints'
Since 'setup.py entrypoints' is using GradleDistribution,
we need to apply 'pythonEnvironmentDistgradle' as the environment.
Otherwise, it couldn't find the project name and version we set as the env variables
in the pygradle plugin.

Currently, we are using 'pythonEnvironment' for it. There won't be any issue if the
product just has a single subproject with the same name as that in product-spec.json.
The build fails, however, if subproject is using a different name OR there are several
subprojects there.

RB=504055
R=sholsapp,broberts,zpetkovi,sfaber,dsully
A=sholsapp