Fix the missing argparse issue for buildPex task.
When the product has no direct dependencies that bring in argparse as a
transitive dependency, then argparse is only an undeclared transitive
dependency through virtualenv->wheel->argparse. This causes the wheel
for it to never be built, which then fails the pex build because it is
in the virtual environment, but no wheel exists for it during packing.

To resolve this:

- add argparse explicitly to build dependencies for Python 2.6 builds
- search through build dependencies (a short list) and if argparse is
  there build a wheel for it with Python 2.6
- this enables the code in pexBuild to allow the argparse dependency to
  get included in the requirements if Python 2.6 is used.

RB=523934
BUGS=TOOLS-66377
R=sholsapp,lcarvalh
A=sholsapp,xachen,dsully