Don't burn in assumptions about transitive deps
Historically, when we create source distributions, we burn in
assumptions about transitive dependencies into the source distribution
itself through a mppy-pinned.txt file. Sometimes, these assumptions blow
up in our face, e.g., when deleting a dependency. What we see is that
all of the burned in mppy-pinned.txt files are outdated. The only way to
update them is to trigger a build.

This change only burns in *direct* dependencies into source
distributions. This satisfies the requirement that our artifacts are
backwards compatible with MPPY and Open Source, and also satisfies the
requirement of not requiring a rebuild when a transitive dependency
changes.

RB=570760
R=zpetkovi,broberts,sfaber,xachen,dsully,psantann
A=zpetkovi,sfaber,psantann,dnozay,ethall,dsully