Python entry point wrapper.
Today, we wrap thin pex files with bash wrappers that invoke pex files
with custom entry points. We want to move this logic slowly to Python
for a number of reasons, mostly around readability and flexibility
moving forward.

In addition to adding a Python entry point, this change sets a custom
PEX_ROOT variable so that pex files unpack themselves to a known
location on disk in QA, CORP and PROD environments. This known location
is created and maintained by SYSOPS. For non-QA, CORP, and PROD
environments we default to use the ~/.pex default.

RB=614272
R=zpetkovi,psantann,ethall,xachen
A=zpetkovi,ethall