Enable choice of the Python interpreter version for build.
This will automatically use our cleanpython interpreters based on the
version selected. It falls back to system Python if the cleanpython is
not present. On Mac, it will fall back to system Python 2.7, if
interpreterVersion = '2.7' was set, but cleanpython27 is not installed
on the machine.

This enables us to start using Python 2.7 immediately.

RB=632691
G=python-foundation-reviewers
R=lcarvalh,ethall
A=psantann,ethall