Fix the repeated rebuilds.
The `pythonMajorMinor` and `pythonVersion` must be updated when
`interpreterVersion` is set. Otherwise, the expected Python version for
eggs and wheels is not correct and they are repeatedly rebuilt.

Making `interpreterVersion` a property with a custom setter solves the
problem. The reason we're leaving `interpreterPath` a custom property is
that it needs to be derived from both `interpreterPath` and
`systemPython` settings. Furthermore, the code quality checker complains
that all of these public fields should be turned into Groovy properties
(private fields with automatic getter/setter) and that all methods are
public by default anyway. I avoided fixing that in this change to keep a
diff focused.

Additionally, I fixed two potential data access issues due to wrong
type:

- a list of numbers was accessed using `[0, 1]` string indexing;
  replaced with `[0..1]` list indexing instead
- the `System.env.PATH` was reported in IDE as unable to resolve;
  replaced with `System.getenv('PATH')`

Other changes that do not affect code:

- added the unit tests that confirm correct behavior for the fixes
- use block comments for long multi-line comments
- use `<p>` between paragraphs in description comments
- do **not** use `<p>` before `@param` because `@` starts the new
  paragraph automatically
- @param, ... description is not captalized and doesn't end in period
- do not use Javadoc comments for private methods; use block comments
  instead

RB=634020
G=python-foundation-reviewers
R=ethall
A=psantann,ethall