Evaluate pythonMajorMinor after interpreterVersion is set.
This will ensure that we process build.gradle first and thus pick up any
new setting of interpreterVersion, before we use pythonMajorMinor to
determine the version used and whether we need to add argparse to the
build configuration. The fix will ensure that 2.7 products such as
wrapped-python-tools doesn't carry an unnecessary argparse dependency in
its Ivy file.

RB=634271
G=python-foundation-reviewers
R=ethall
A=psantann,ethall