Update pex, setuptools, and friends.
We want to update setuptools for a long time now.
Due to the pex update we also update virtualenv, wheel, and argparse as
a side effect.

Refactored to pin in a single place (a mapping) instead of using version
numbers all over the code base.

Also, use the new distgradle that works better with newer setuptools.

RB=694631
G=python-foundation-reviewers
R=tgoetze
A=sholsapp,xachen