Revert "Revert "Update setuptools, pex, and others.""
This reverts commit 6f8cf40fda98a43f2e94807b10965ad5942aebd3.

After the revert, refactored to get the configuration mappings into
PINNED_VERSIONS map. This enables us to avoid using the concatenated
strings and improves the readability.

Added the new input for PipInstallTask: `sorted`.
By default, it's `true` to install the dependencies in sorted order.
It can be reset to `false` to enable custom ordering for bootstrap/setup
requirements.

Created a new configuration, specifically for setup requirements, that
is installed in the order needed for bootstrapping.

Incremented the minor version of the product to 0.1.0

RB=703491
BUG=TOOLS-105462
G=python-foundation-reviewers
R=ethall
A=sholsapp,psantann,ethall