Fix the test configuration dependencies.
The test configuration dependencies can overwrite the python
configuration dependencies if they resolve to a different version
transitively. By extending from the python configuration, we ensure that
this doesn't happen and also do the semantically correct dependency
declaration.

Later when we do `pipFreeze`, we take care to remove all dependencies
that are in the python configuration, which we were doing anyway in an
`if` statement before, but this time we do it more explicitly using only
one set.

RB=714605
BUG=TOOLS-105828
G=python-foundation-reviewers
R=ethall
A=sholsapp,dnozay,ethall