Update Sphinx and flake8 and fix the flake8 taks order.
This updates Sphinx to 1.4 and makes it a build dependency.
This is correct, because it produces the docs artifacts.

Also, flake8 and its dependencies are updated. This includes the fix for
pep8 warning about Python 3 compatibility.

Finally, the order of the flake8 taks is changed to appear much earlier
in build. This will enable our precommit hooks to finish much faster then
they do now.

RB=716566
BUG=TOOLS-87056,TOOLS-96689
G=python-foundation-reviewers
R=sholsapp,tmcginni,ethall
A=ethall,dsully