Adding a pinned.txt to projects so Pip can work outside of pygradle
This adds a pinned.txt to the project to automatically be added to the
project under the projectDir. This also adds a test to verify that it
works as intended.

This also adds a task where users can have a setup.py generated into
their project so they don't have to remember what to copy/paste into
the repo.