Making the consule output contain test information
This change makes it so we will log the test results to the consule and
when running in CI mode we will print the results of teach test as they
finish.

This also makes it so that unit tests run before integ tests so that we
are more likely to find failures on the quick tests vs finding them on
the long tests.