Moving from using project to setting the deletage on the closure
This should also make sure that if people have pygradle loaded their
groovy IDE should be able to let you know what the options are for the
closure.

This removes an integ test and adds a unit test. The integ test sets a
value to make sure the parse works, but doesn't require another test
(which adds a bunch of time to builds)