Fixing pip install where it wouldn't install properly
It looks like we were setting the pythonEnvironmentDistgradle into all
pip install tasks, which is a bad because when we install a project that
uses the env, we'll overwrite the data and cause the package to install
overplace the current project.

This remove that problem. The local-install that happens does add the
pythonEnvironmentDistgradle to the env by using the environment property
to the task, so the installs to the project work properly

Adding an ivy repo so we can publish to Ivy too.