Adding better support for thin pex templates
Before we would only accept a string, but that's not useful if someone
needs to pick different options. This way the plugin will now use the
same thing that a user might want to change it to. Also since it's a SAM
it's easy to write a closure for it.