Fix the handling of special cases.
The Python packages with underscores in their name have the tar files
with the underscore ("_") in the file name, but the dash ("-") in the
package name in the metadata. Similarly in Ivy files for Gradle.
Instead of saving each special case in the map, we're going to handle it
generically for each such possible package name.