Fix the coverage task.
The arguments for coverage are inserted between `py.test` and the test
directory. Since we moved the adding of `py.test` to exec args from
constructor to preExecution, we now need to ensure this insertion of
extra arguments for coverage in a different way.

Temporarily disable findbugs to fit into CI memory limit.