Finishing up tests to make them work
Things left:
- Make the tests actually work. Right now it looks like the exec can
run, but windows doesn't honor the shebang line so we have to make the
pex's end with .py so they can run :-(
- Drinking.

This change has a lot in it, mostly because removed the
VirtualEnvExecutableHelper file and replaced it with the
VirtualEnvironment. The former required a PythonDetails to do anything,
so instead of doing that now you just call into details and ask for the
venv. All of the methods that were static no loger are, and life is
better :-)