Change installation order to post-order graph traversal.
Add the new utility that resolves the dependencies in the post-order.
Use it to get a file collection and apply that collection to the
PipInstallTask and BuildWheelsTask for all configurations.

Now we can drop pbr from setup requirements.

Added comprehensive unit tests for the utility class.

Added additional assertions in an existing integration test.
Checked that it passed with expectation for sorted build dependencies on
master and that it failed with the same expectation with this change.
Once the condition was negated, it passed with this change.

Bumped the version to 0.5.0.