Prepare for upgrade of build and setup requirements.
Pex has changed the default behavior and does not allow pre-release
packages by default. That's a very noble goal, but it unfortunately
breaks a lot of builds. We are providing a task input (optional) that
allows us to pass additional options to the pex build task.

Pip has changed the format for the freeze. It now returns the currently
developed package as "-e repo-url-here" instead of "name==version". We
had to adapt to that change and put the current package name
unconditionally in.

Both changes were implemented in a way that will work with old or new
versions which opens a path for upgrades.