Fix pivy-importer.
The importer didn't handle conditions with not equal, greater than, or
less than correctly. It also didn't look into the ranges of
dependencies. This fixes the parsing of conditions, and provides for
avoiding the pre-releases (default) beside sorting them correctly below
stable releases. It also adds the option to use latest versions of
dependencies instead of lowest, which is important for the case when a
requirement is greater than some very old release.