Remove duplicate code path for building pex.
Use only the new generator code.
The old utility functions are removed and their calls refactored.
Tested on a build of a web app, with or without fat pex.

Fixed the bug in the regex and match group used in the new
PexExecOutputParser, based on the correct code in the old utility
function. Tested on a real build with an artifically introduced error
that the error report contains only the package name, instead of the
whole match sentence.