Upgrade to pex-1.2.13 and six to 1.11.0.
This version of pex contains our fix for handling `--pre` option
correctly. The new version of six has improved metaclass compatibility.

We also force the version of wheel, besides setuptools, in pivy-importer
because of pex's tight requirements over these two dependencies.

The test had to be fixed to include product version now that we pass it
to pex command.