Converted classes from Groovy to Java.
** Note **
We are dropping Java 7 support. We are going to use Java 8 features to
make the code base more useable and easiser to manage

In an effort to modernize the codebase and allow innerop with Kotlin

This also splits the base PythonPlugin into multiple componenets. This
is done to make the code more readable, and make parts of the code base
live closer together. The split up "plugins" don't have a public ID so
they are intended to be a private implementation detail.