Fixed bug in WheelCache in picking wheels
There was a bug in the wheel cache in which the file names were used and
there was a string match, but the match didn't have all the parts.

Now we'll break the wheel up into parts that match the wheel file
format. Then use the fields from the file format to filter the wheels.