First pass at updating the PyPi Parser
We've seen issues with the Pivy Importer where as the world moves
forward our tests fail. Since we really want something that's finalized,
to isolate ourselves for tests, this change makes it so that you're
required to supply a specific version.

The version input is in "requires.txt" format. This makes it so that a
normal python project could use it as input.

This is just a first pass, but we should be able to easily extend it to
do proper version validation.