Adapt integration test regex to travis.
Travis injects the commit message into the environment as a variable.
This may result in a new line being injected if the commit message
contains a new line, which is the case most of the time. In that case
the existing regular expression does not work. Instead we're going to
capture the whole environment between the braces {...} and look for the
variables we're interested in inside the captured match. Also, use /.../
instead of '...' for all regex parameters to find().