Parallel Wheels will show all wheels being built
Making the user more aware of which wheels are currently being built, by
showing it to the user.

We're also reducing the number of threads avaliable to do the parallel
building incase gradle is trying to do something else.

I've noticed by dropping the parallelization, we get some slightly
faster builds (30 seconds on a 10 min build).