Created DistUtilsVirtualEnvCustomizer and extracted Wheel builds
Since we need the ability to create venvs that may be customized, we
should make tasks that create a venv support that as a first class way.
This is done via the 'SupportsDistutilsCfg' interface. Both the venv
task and the flake8 task implement it.

The logic that was in Venv task was extracted into DistUtilsVirtualEnvCustomizer
and that is now always passed. Keeping the null check for now, we can
remove it in the future.

In an effort to fully replace the integ test, this also creates a
`PipWheelAction` class that is like `PipInstallAction`. The wheel action
allows us to test the install steps without actually having to do an
install. Keeping the tests quick, and remove some flakeyness in the
testing.

Added an override method on PackageInfo to accept Paths, it just
delegates to the File method. This is useful for tests.

Fixed the tests where they would fail on Windows because we were using
unix style paths instad of Windows style paths.