Extracted Interfaces from Classes, and moved static methods
Removed `new PythonDetails` and replaced it with a factory call to
`PythonDetailsFactory.makePythonDetails`.

The PythonDetails object was replaced with an interface. This can allow
us in the future to "shadow" some values when needed (like with infra
venvs).

Also created an interface for VirtualEnvironment and moved the impl into
internal.

The PythonDetailsFactory now supports a "clone" option, so we can re-use
most of a PythonDetails but replace the venv directory. This is used for
infra tasks. We want to use the main PythonDetails for everything,
except the venv.

Moved code from Flake8Task into the
AbstractPythonInfrastructureDefaultTask class.