ParallelWheels needed SetupRequires before building wheels
There was an issue where the ParallelWheels task wouldn't have the
setupRequires dependencies installed until after the wheel was
generated. This causes issues with namespace pacakges and can lead to
indeterminate import errors.

Also, we should not build wheels and put them into the global cache
with extra dependencies other than the setup-requires deps. So we check
to see that there are only the dependenices we care about in the venv.