Refactored IntegTest to become unit test. (#237)
* Refactored IntegTest to become unit test.

PackageSettingsIntegrationTest was flakey and really just wanted to
ensure that specific arguments were passed to pip. So now of running a
build, we will validate that the arguments would have been passed to
pip. This makes the tests fast and not rely on the env. It also means
that we don't need to do any strange regex's and rely on side affects of
the build processes to validate things.

The previous test was asserting that Gradle was passing args to the
commandline, and that's the same as asserting that our code was calling
Gradle with the correct parameters.

This change splits the integ test into two tests, because it was testing
two different things. `pip install` and `pip wheel`, so now there is one
test for each.

To do this, two actions were extracted. PipInstall and PipWheel. They
both use the same base class so we can re-use as much code as possible.

* Addressing Zvezdan's comments

- Deleted unused class
- Fixed some package declerations
- Using PipWheelAction in BuildWheelTask

* Updated tests to assert rebuild of wheels.

Other misc changes to codebase based on feedback

* Moving code to Java

Noticed that this was all java, so removing the Groovy plugin and the
classes are moved into the Java sourceSet

* Added more test coverage

Tests now verity skipping behavior

* Test location matches package definition

* Fixed FailureReasonProvider for BuildWheels

This change adds an integ test a few tasks that also use
FailureReasonProvider.

AbstractPythonMainSourceDefaultTask was updated to always add details to
the result call.