Make enabling coverage a little easier
Taking a cue from the recent mypy extension, add CoverageExtension class with
a `run` property.  Make the check task depend on the coverage task so that it
is run automatically, but only if `run` is true.  Now I can simplify running
coverage by just adding this to my build.gradle:

```
python {
  coverage.run = true
}
```

Closes #182