Refactor Python plugins for more extensible container formats
Remove the hard code to PexExtension

* Restore PythonPexDistributionPlugin.java
* Implement a default fallback container extension; we still need a config option

addBuildTask() -> addTasks()

Be sure to only build the wheels once

Use a TaskContainer local variable for readability

The Tar task doesn't need to use a closure

Use a lifecycle task to handle dependency extensions

* buildContainers is the lifecycle task name
* PythonContainerPlugin now just has a makeTasks() API.  Extensions are
  expected to use this and *not* tie their task into the dependency
  hierarchy.  PCP will do that by default.

Rework the tasks so Tar is only added once

buildContainers -> assembleContainers

Add build.gradle ui for setting the container extension

Refactoring based on comments

* PexExtension -> PexApplication
  - Fix up interfaces
  - Moved to tasks
  - Removed ExtensionUtils.maybeCreatePexExtension() since this isn't
    technically a Gradle extension
* ContainerExtension -> ApplicationContainer
  - Moved to tasks
* ZipappContainer extends ApplicationContainer
* Instance variable renames
* Code style cleanup

Move the container classes to the util package

Extensions must be added before build.gradle is evaluated

...otherwise, there's no e.g. `python.pex` section to add configuration to.
Similarly, dependencies and tasks must be added *after* build.gradle
evaluation, so they must be moved to an `afterEvaluate` closure.

Also:

* Add ApplicationContainer.addExtension() API
* Move PexApplication to util package

Move PythonContainerTasks to tasks package

We have to get the new application container in the afterEvaluate()

Change the way the default container is returned

* If container is null, return the default application container
* If it is not null, but unknown, then null is returned

Then, in the PythonContainerPlugin, we check the application container, and if
it is null, we know the build.gradle requested an unknown container type, and
we can raise an exception.

Refactor entry point template provider for web apps

Canonicalize comments

Add TODO comment as requested by review

PexApplication -> PexExtension