Fix the handling of generated code and store overwrite.
A generated code, such as rest.li, is also being installed from a
directory, not an sdist package. We need to handle the package version
number similar to the project directory.

Two sub-projects building at the same time can cause the behavior where
both of them do not find a wheel in any layer and proceed to build
it in their project layers. Then one of them stores it into the host
layer before another. The second one should not fail if that wheel is
present when it tries to store it in the host layer.

Added tests for both cases and a safety belt against null.