Treat project wheel as a special case.
The project wheel will be built after the editable install for
development. This ensures that the file manifest is generated for the
project package before the wheel is built. Otherwise, wheel does not
include resource files that `include_package_data=True` ensures to be
packed into an sdist.

This also helps some packages that generate code after the development
install and expect that code to be packed into the wheel.