Add a safety belt for wheel readiness.
When a user removes wheels or wheel cache in the project layer, but does
not clean up virtualenv, the wheel will not be present for the packaging
of the zipapp later. This safety belt ensures that we drop the flag for
wheel readiness in that case and thus allow build wheels task later to
run and build all missing wheels.

Also, added the tests for all the "impossible" code paths by hooking in
the appropriate fake objects and triggering the conditions in
safety-belt statements.