Refactor MypyTask and MypyExtension for local customizations
If and when https://github.com/python/mypy/pull/7219 lands, we will have the
ability to tell mypy to traverse into PEP 420 namespace packages.  In order to
preserve backward compatibility, the MypyExtension now allows clients to set
the mypy command line arguments, but if they don't it will use the old command
line of just passing the source directory.