Modified ValidationPlugin two ensure pytests only run once
Currently, if coverage is enabled for a build, the pytest
task is run and then the coverage task is run, which inherits
from the pytest task, and consequently, the pytest suite is
run twice. Added logic to ensure the either pytest XOR coverage
task is part of the build. Also added logic to existing
integration tests (PythonPluginIntegrationTest, PexIntegrationTest)
to verify that the coverage/pytest tasks are skipped when expected