Added integration tests, made changes based on comments
Realized that if coverage fails, it has the process return
a non-zero return value, meaning it will fail the build. This
is actually desired behavoir, but I didn't realize I could
accomplish this is the same PR. Added integration tests in
PythonPluginIntegrationTest.groovy to make sure the build fails
as expected if coverage fails (or if coverage succeeds but pytest
fails). Also modified the logic in ValidationPlugin.java to
use a shared BiPredicate between the pytest and coverage task
that is used to determine if they should run. Also added a
logging statement if the pytest task is being skipped.