Reverted logic that has flake8 run in different venv
Modified the Flake8Task to once again inherit from
AbstractPythonMainSourceDefaultTask and use the original
virtual environment instead of having a dedicated venv just for
it's run. Deleted the AbstractPythonInfrastructureDefaultTask as
Flake8Task was the only class using it.

Modified PythonPlugin to have flake8 be added as a build configuration
instead of it's own configuration. Deleted unit and integration tests
ensuring that flake8 was run in its own environment.