Added a null check for the htmlDir in PyCoverageTask
Handling of an edge case: If the user configures the project
to run covergae but there is no coverage output (i.e. no
coverage to report), the htmlDir parsed from the output will
ultimately be null, and thus the call to project.file(htmlDir)
fails with IllegalArgumentException: path may not be null or
empty string. path='null'