Upgrade flake8 to 3.7.9.
Have to use no-build-isolation option for pip on a few integration tests
because flake8-3.7.x depends on entrypoints that require flit in a
build_system section of the toml file. We don't need that since we
already have a package with flit-built setup.py. In future release of
flake8 entrypoints are replaced with importlib-metadata, so this
dependency will not cause us issues any more, but the no-build-isolation
setting should be fine in the tests even in that case.